+++ bin/nanocld/migrations/2024-01-02-135246_metrics_rename_expire/up.sql
@@ -1,2 +1,0 @@
--- Your SQL goes here
-ALTER TABLE IF EXISTS "metrics" RENAME COLUMN "expire_at" TO "expires_at";