+++ bin/nanocld/changelog.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.15.0] - 2024-05-16
+## [0.15.0] - 2024-06-11
 
 ### Chore
 
@@ -24,13 +24,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Filter query by dates
 - Updating a secret now trigger an update for related cargoes
 
-
 ### Fixed
 
 - Removing a job with a running process will now stop the process before removing the job
 - Patch a restarting cargo was not working and returning an error
 - Create network namespace if not existing when inspecting a listing namespaces
-- Patching cargo when restarting was not working
 
 ### Changed